Add support for Smart Contract Signatures in transactions
