+++ packages/protocol-kit/src/utils/signatures/utils.ts
@@ -131,3 +131,42 @@ export async function generateEIP712Signature(
 
   return new EthSafeSignature(signerAddress, signature, true)
 }
+
+export const buildSignatureBytes = (signatures: SafeSignature[]): string => {
+  const SIGNATURE_LENGTH_BYTES = 65
+
+  signatures.sort((left, right) =>
+    left.signer.toLowerCase().localeCompare(right.signer.toLowerCase())
+  )
+
+  let signatureBytes = '0x'
+  let dynamicBytes = ''
+
+  for (const sig of signatures) {
+    if (sig.isSmartContractSignature) {
+      /* 
+        A contract signature has a static part of 65 bytes and the dynamic part that needs to be appended 
+        at the end of signature bytes.
+        The signature format is
+        Signature type == 0
+        Constant part: 65 bytes
+        {32-bytes signature verifier}{32-bytes dynamic data position}{1-byte signature type}
+        Dynamic part (solidity bytes): 32 bytes + signature data length
+        {32-bytes signature length}{bytes signature data}
+      */
+      const dynamicPartPosition = (
+        signatures.length * SIGNATURE_LENGTH_BYTES +
+        dynamicBytes.length / 2
+      )
+        .toString(16)
+        .padStart(64, '0')
+
+      signatureBytes += sig.staticPart(dynamicPartPosition)
+      dynamicBytes += sig.dynamicPart()
+    } else {
+      signatureBytes += sig.data.slice(2)
+    }
+  }
+
+  return signatureBytes + dynamicBytes
+}
