+++ packages/protocol-kit/src/utils/transactions/SafeTransaction.ts
@@ -3,6 +3,7 @@ import {
   SafeTransaction,
   SafeTransactionData
 } from '@safe-global/safe-core-sdk-types'
+import { buildSignatureBytes } from '../signatures'
 
 class EthSafeTransaction implements SafeTransaction {
   data: SafeTransactionData
@@ -17,16 +18,7 @@ class EthSafeTransaction implements SafeTransaction {
   }
 
   encodedSignatures(): string {
-    const signers = Array.from(this.signatures.keys()).sort()
-    const baseOffset = signers.length * 65
-    let staticParts = ''
-    let dynamicParts = ''
-    signers.forEach((signerAddress) => {
-      const signature = this.signatures.get(signerAddress)
-      staticParts += signature?.staticPart(/*baseOffset + dynamicParts.length / 2*/).slice(2)
-      dynamicParts += signature?.dynamicPart()
-    })
-    return '0x' + staticParts + dynamicParts
+    return buildSignatureBytes(Array.from(this.signatures.values()))
   }
 }
